refactor: rename hooks.ts to rendering.ts
